little reog of new function to split into processes
