+++ src/initialize_program.c
@@ -12,6 +12,17 @@
 
 #include "../include/minishell.h"
 
+void	sig_handler(int signum)
+{
+	if (signum == SIGINT)
+	{
+		ft_putstr_fd("\n", 1);
+		rl_on_new_line();
+		rl_replace_line("", 0);
+		rl_redisplay();
+	}
+}
+
 void	initialize_signals(void)
 {
 	struct sigaction	action_1;
